requirements.txt
numpy >= 1.2.0
h5py >= 1.0.0
netcdf4 >= 1.0.0
tqdm >= 4.0.0
pyshp >= 2.1.0
pillow >= 6.0.0
imageio >= 2.15.0
sphinx-rtd-theme >= 1.3
shapely
